+++ src/utils/test.ts
@@ -242,6 +242,9 @@ export const createPanelOptions = (options: Partial<PanelOptions> = {}): PanelOp
   tabsSorting: false,
   toolbar: createToolbarOptions({}),
   nestedObjects: [],
+  isColumnMangerAvailable: false,
+  saveUserPreference: false,
+  showFiltersInColumnManager: false,
   ...options,
 });